Adding Code for Forest_gathering_problem
Chef is the head of commercial logging industry that recently bought a farm containing N trees. You are given initial

height of the i-th tree by H i  and the rate of growth of height as R i  meters per month. For simplicity, you can assume

that all the trees are perfect cylinders of equal radius. This allows us to consider only the height of trees when we talk

about the amount of wood.

In Chef&#39;s country, laws don&#39;t allow one to cut a tree partially, so one has to cut the tree completely for gathering wood.

Also, laws prohibit cutting trees of heights (strictly) lower than L meters.

Today Chef received an order of W meters (of height) of wood. Chef wants to deliver this order as soon as possible.

Find out how minimum number of months he should wait after which he will able to fulfill the order. You can assume

that Chef&#39;s company&#39;s sawing machines are very efficient and take negligible amount of time to cut the trees.

Input

There is a single test case per test file.

The first line of the input contains three space separated integers N, W and L denoting the number of trees in the

farm, the amount of wood (in meters) that have to be gathered and the minimum allowed height of the tree to cut.

Each of next N lines contain two space separated integers denoting H i  and R i  respectively.

Output

Output a single integer denoting the number of months that have to pass before Chef will be able to fulfill the order.

Constraints

 1 ≤ N ≤ 10^5

 1 ≤ W, L ≤ 10^18

 1 ≤ H i , R i  ≤ 10^9

Subtasks

 Subtask #1 [40 points]: 1 ≤ N, W, L ≤ 10^4

 Subtask #2 [60 points]: No additional constraints

Input:

3 74 51

2 2

5 7

2 9

Output:

7

Explanation

After 6 months, heights of each tree will be 14, 47 and 56 respectively. Chef is allowed to cut only the third tree, sadly

it is not enough to fulfill an order of 74 meters of wood.

After 7 months, heights of each tree will be 16, 54 and 65 respectively. Now Chef is allowed to cut second and third

trees. Cutting both of them would provide him 119 meters of wood, which is enough to fulfill the order.